fix(TDC) Make ConditionChecker a RegisteredClass
Storages can have mandatory condition checkers defined, so register that class
so it can be exposed automatically in configuration.
