+++ snuba/datasets/storages/outcomes.py
@@ -7,7 +7,7 @@ from snuba.datasets.schemas.tables import TableSchema, WritableTableSchema
 from snuba.datasets.storage import ReadableTableStorage, WritableTableStorage
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.datasets.table_storage import build_kafka_stream_loader_from_settings
-from snuba.query.processors.physical.conditions_enforcer import OrgIdEnforcer
+from snuba.query.processors.condition_checkers import OrgIdEnforcer
 from snuba.query.processors.physical.prewhere import PrewhereProcessor
 from snuba.query.processors.physical.table_rate_limit import TableRateLimit
 from snuba.utils.streams.topics import Topic
